Conflicts:
	Aufgabe1/src/gkaAufgabe1/.classpath
	Aufgabe1/src/gkaAufgabe1/bin/.gitignore

+++ Aufgabe1/src/gkaAufgabe1/src/Gui/WeightedEdge.java
@@ -10,7 +10,7 @@ public class WeightedEdge extends DefaultWeightedEdge {
 		super();
 	}
 	
-	public WeightedEdge(int weight) {
+	public WeightedEdge(double weight) {
 		
 	}
